feat, chore, docs, refactor: Modifications in the 'db_manager.py' and test_BirDayBer.py files, including .gitignore.
refactor --> Change in the naming in the BirDayBerDB_manager.py:
db_manager.py | class Birthdays_db() --> Db_manager() | method
add_person() --> add_attributes()

feat --> Change in the internal functionality of the  add_attributes
method. Now, it return a list with tuples to process the data in a new
method of the same class that doesn't exist yet, called
'process_attributes'.

docs --> With the refactor, the names in the docs of the testing and db_manager.py files have changed too.

chore --> Modification in the .gitignore file.