fix, refactor, test: Modification in the db_manager.py and test_birdayber.py files.
fix & test --> Bug fix in the test_BirDayBer.py file, where the
'test_check_ages' method was expecting something that i shouldn't.

test --> And change in the method of the tests inside the class
BirthDB_testing. Now the db_manager.py it's not supposed to use multiple
methods to fulffil the comprobation of the values, now it's just one
method. This one is 'column_search' and it will get the values to search in the parameters.

refactor --> In the naming of olmost every method in the db_manager.py file.