fix, test: Bug fix in the db_manager.py file and adaptation of the test.
fix --> The 'column_search' method in db_manager.py wasn't able to show
more than 1 column per time (When this was not it's original target),
but olso when you tries, you will only get the first column (generally
the ID).

test --> The test now is adapted to the new output of the 'column_search'
method.