feat & fix & refactor: Refactor in the test_BirDayBer.py file, fix in the db_manager.py but olso feat in this file too.
refactor --> In the setUp method of the test_BirDayBer.py file.

fix & feat --> In the 'process_rows()' method of the db_manager.py
file. There is a bug that didn't allow the user to use 'Inner join',
especifically a logical mistake by me. Now, it's all prepared to treat
this problem.