fix, chore, test: Little modification to the README.MD file and progress in the fix of the join clauses.
chore --> Change of text positions in the README.MD file.

fix --> Advance in the fix of the Join clause in the db_manager.py file,
now it doesn't give an empty list, it gives the last row.

test --> Improve in the test 'test_check_data' of the test_BirDayBer.py
file.