feat, test, chore: First steps to implement images in the database.
feat --> Implementation of the 'photo_to_binary' function and
modifications in the 'process_rows' method in the db_manager.py file.

test --> Modification of the first 'photo' value of the 'Photo' table. That value is a None no longer, now is an image direction.

chore --> Implementation of an experimental image in the 'conceptual
folder'