feat, chore, test, docs: Implementation of the function 'binary_to_photo' in db_manager.py file and new folders.
feat --> Implementation of the function 'binary_to_photo' in db_manager.py, and
with this, there are new folders that works with this function.

chore --> Implementation of several folders to keep the general program folder clean.

test --> Adaptation of the 'test_check_photos' test in the
test_BirDayBer.py file for the new function in the db_manager.py.

docs --> Addition of new documentation for the new function.