feat, chore, test, docs: Modification of the whole structure of the folders, implementation of the drop_database method and modification of the README.md file.
feat --> Implementation of the 'drop_database' method in the
db_manager.py file.

chore --> Modification of the structure of the build and update of the
README.md file. And the test_db.db file now is save in the 'data/database/' folder.

test --> Update of the 'test_check_photos' in test_BirDayBer.py file.
And preparations to check if the 'column_search' function can create
certain photos in the 'data/rows_content/' folder.

docs --> Update of some method in the db_manager.py file.