feat, test, docs: First advance in the BirDayBer.py file and some modifications of the test_BirDayBerClient.py testing file.
feat --> Implementation of the 'Birdayber_client' class, with the
'window_resolution' method included in the BirDayBer.py file.

docs --> Implementation of documentation to the 'Birdayber_client' class
and 'window_resolution' method in the BirDayBer.py file.

test --> Few modifications, implying the current uselessness of
the 'setUp' and 'tearDown' methods in the test file 'test_BirdayBer Client.py'.