feat, chore, test: Elimination of 1 test, implementation of the 'get_license' method and little modification of the README.md file.
feat --> Implementation of the 'get_license' method in the BirDayBer.py
file.

chore --> Little modification at the end of the file in the README.md
file, changing the text of the 'MIT LICENSE' to 'MIT © Luciano Esteban'.

test --> Elimination of the test 'test_birthdayNotation' of the
test_BirdayberClient.py testing file.