feat: Modification in the structure of the BirDayBer.py file code, dividing the 'Birdayber_client' class in two.
feat --> Now the Birdayber.py file has two classes to work on.
The first one is the father class 'Birdayber_system', that is in charge
to generate and work with the database.
The second one is the child class 'Birdayber_client', that is in charge
to generate and work with the visual interface.