test, chore: Change in the 'test_image_generator' of the BirDayBer testing file and the structure of the folders.
test --> Change in the naming of the 'test_image_generator' test -> 'test_default_images'.
Also the content of the test has been reduced.

chore --> Implementation of the 'Responsive' folder in the
'BirDayBer\bin\system_content\visual_content\responsive' path.
Also elimination of the '.keep.' file on the 'visual_content' folder