chore, feat, test: Modificaton of the client's test and implementation of the beginning of the 'responsive_imgs' method in the BirDayBer.py file
feat --> Implementation of the 'responsive_imgs' method in the
'Birdayber_client' class of the BirDayBer.py file (still in process).

chore --> Modification in the README.md file, mentioning the use of a
new library in the program.

test --> Modification of the 'test_responsive_images' and
'test_default_images' of the 'test_BirDayBer.py' testing file.